+++ benchmarks/dom-comparison/src/implementations/react-fela-useFela/Provider.js
@@ -1,7 +1,7 @@
 /* eslint-disable react/prop-types */
 import React from 'react'
 import { createRenderer } from 'fela'
-import { RendererProvider as FelaProvider } from 'react-fela'
+import { RendererProvider  } from 'react-fela'
 import View from './View'
 
 const renderer = createRenderer()
@@ -9,9 +9,9 @@ const renderer = createRenderer()
 class Provider extends React.Component {
   render() {
     return (
-      <FelaProvider renderer={renderer}>
+      <RendererProvider renderer={renderer}>
         <View>{this.props.children}</View>
-      </FelaProvider>
+      </RendererProvider>
     )
   }
 }