+++ packages/fela-monolithic/src/__tests__/monolithic-test.js
@@ -18,6 +18,18 @@ describe('Monolithic enhancer', () => {
     expect(renderer.cache.hasOwnProperty(`.${className}`)).toEqual(true)
   })
 
+  it('should add extra _className', () => {
+    const rule = () => ({
+      _className: 'foo',
+      color: 'red',
+    })
+    const renderer = createRenderer(options)
+
+    const className = renderer.renderRule(rule)
+
+    expect(className).toContain('foo')
+  })
+
   it('should reuse classNames', () => {
     const rule = props => ({
       color: props.color,
@@ -89,6 +101,24 @@ describe('Monolithic enhancer', () => {
       `.${className}{color:red}.${className}:hover{color:blue}`
     )
   })
+  it('should render pseudo classes, with specificityPrefix', () => {
+    const rule = () => ({
+      color: 'red',
+      ':hover': {
+        color: 'blue',
+      },
+    })
+
+    const renderer = createRenderer({
+      ...options,
+      specificityPrefix: '.parent ',
+    })
+    const className = renderer.renderRule(rule)
+
+    expect(renderToString(renderer)).toEqual(
+      `.parent .${className}{color:red}.parent .${className}:hover{color:blue}`
+    )
+  })
 
   it('should prefix classNames', () => {
     const rule = () => ({
@@ -101,6 +131,22 @@ describe('Monolithic enhancer', () => {
     const className = renderer.renderRule(rule)
 
     expect(renderToString(renderer)).toEqual(`.${className}{color:red}`)
+    expect(className).toContain('fela_')
+  })
+  it('should prefix classNames, and prefix selectors with specificityPrefix', () => {
+    const rule = () => ({
+      color: 'red',
+    })
+
+    const renderer = createRenderer({
+      selectorPrefix: 'fela_',
+      specificityPrefix: '[class|="fela_"]',
+    })
+    const className = renderer.renderRule(rule)
+
+    expect(renderToString(renderer)).toEqual(
+      `[class|="fela_"].${className}{color:red}`
+    )
     expect(className).toContain('fela_')
   })
 
@@ -133,6 +179,24 @@ describe('Monolithic enhancer', () => {
 
     expect(renderToString(renderer)).toEqual(
       `.${className}>div{color:blue}.${className}{color:red}`
+    )
+  })
+  it('should render child selectors, and prefix specificityPrefix', () => {
+    const rule = () => ({
+      color: 'red',
+      '>div': {
+        color: 'blue',
+      },
+    })
+    const renderer = createRenderer({
+      ...options,
+      specificityPrefix: '#app ',
+    })
+
+    const className = renderer.renderRule(rule)
+
+    expect(renderToString(renderer)).toEqual(
+      `#app .${className}>div{color:blue}#app .${className}{color:red}`
     )
   })