+++ packages/fela-plugin-fullscreen-prefixer/src/index.js
@@ -1,0 +1,14 @@
+/* @flow */
+import pseudoPrefixer from 'fela-plugin-pseudo-prefixer'
+
+const fullscreenPrefixes = [
+  ':-webkit-full-screen',
+  ':-moz-full-screen',
+  ':-ms-fullscreen',
+  ':full-screen',
+  ':fullscreen',
+]
+
+export default function fullscreenPrefixer() {
+  return pseudoPrefixer(':fullscreen', fullscreenPrefixes)
+}