+++ CHANGELOG.md
@@ -5,6 +5,28 @@ If you're searching for older logs, please check the [old changelog](https://git
 
 ## 11.0
 
+### 11.4.0
+| Package | Changes |
+| --- | --- |
+| fela-plugin-theme-value | ([#812](https://github.com/robinweser/fela/pull/812)) Introducing a new plugin to access theme values directly |
+| fela-plugin-pseudo-prefixer<br>fela-plugin-fullscreen-prefixer | ([#814](https://github.com/robinweser/fela/pull/814)) Introducing 2 new plugins: a general pseudo prefixer and a specific `::fullscreen` prefixer built on top of that |
+| fela<br>fela-dom | ([#813](https://github.com/robinweser/fela/pull/813)) Added a new renderer config option called `specificityPrefix` that allows to prefix generated selectors with a global prefix |
+| fela-dom | ([#811](https://github.com/robinweser/fela/pull/811)) Renamed UMD output to `FelaDOM` to match previous versions |
+
+### 11.3.2
+
+| Package | Changes |
+| --- | --- |
+| fela-monolithic | Added the ability to pass a _className that's added to the monolithic className output |
+| fela-dom | ([#805](https://github.com/robinweser/fela/pull/805)) Added an UMD bundle |
+
+### 11.3.1
+
+| Package | Changes |
+| --- | --- |
+| fela | ([#795](https://github.com/robinweser/fela/pull/795)) Removes the warning when fonts are rendered with a query string attached |
+| react-fela<br>preact-fela<br>jest-react-fela | ([#792](https://github.com/robinweser/fela/pull/792)) Added typings for `fe` and `createSnapshot`<br>([#796](https://github.com/robinweser/fela/pull/796)) Added typings for `rehydrate` and `targetDocument` props on `<Provider>` |
+
 ### 11.3.0
 
 | Package | Changes |